requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
idna==3.10
    # via requests
praw==7.8.1
    # via -r requirements.in
prawcore==2.4.0
    # via praw
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via
    #   prawcore
    #   update-checker
tabulate==0.9.0
    # via -r requirements.in
update-checker==0.18.0
    # via praw
urllib3==2.2.3
    # via requests
websocket-client==1.8.0
    # via praw
